migrate to float

+++ effects_engine.h
@@ -19,17 +19,17 @@ typedef CRGB Pixel;
 typedef std::vector<Vector> VertexList;
 typedef std::vector<std::vector<unsigned int>> AdjacencyList;
 
-typedef std::function<Pixel(const Vector&, double)> ColorFn;
-typedef std::function<Pixel(double x, double y, double t)> TrailFn;
+typedef std::function<Pixel(const Vector&, float)> ColorFn;
+typedef std::function<Pixel(float x, float y, float t)> TrailFn;
 
-typedef std::function<Vector(double)> PlotFn;
+typedef std::function<Vector(float)> PlotFn;
 
-typedef std::function<double(double)> EasingFn;
-typedef std::function<double(double)> ShiftFn;
-typedef std::function<double(double)> MutateFn;
-typedef std::function<double(double)> WaveFn;
+typedef std::function<float(float)> EasingFn;
+typedef std::function<float(float)> ShiftFn;
+typedef std::function<float(float)> MutateFn;
+typedef std::function<float(float)> WaveFn;
 
-typedef std::function<void(Canvas&, double)> SpriteFn;
+typedef std::function<void(Canvas&, float)> SpriteFn;
 typedef std::function<void(Canvas&)> TimerFn;
 
 #include "geometry.h"
