+++ filter.h
@@ -17,8 +17,8 @@ public:
     return filter;
   }
 
-  virtual void plot(Canvas& canvas, double x, double y,
-    const Pixel& c, double age, double alpha) = 0;
+  virtual void plot(Canvas& canvas, float x, float y,
+    const Pixel& c, float age, float alpha) = 0;
 
   virtual void decay() {
     if (next != nullptr) {
@@ -26,7 +26,7 @@ public:
     }
   }
 
-  virtual void trail(Canvas& canvas, TrailFn trailFn, double alpha) {
+  virtual void trail(Canvas& canvas, TrailFn trailFn, float alpha) {
     if (next != nullptr) {
       next->trail(canvas, trailFn, alpha);
     }
@@ -34,8 +34,8 @@ public:
 
 protected:
 
-  void pass(Canvas& canvas, double x, double y,
-    const Pixel& c, double age, double alpha)
+  void pass(Canvas& canvas, float x, float y,
+    const Pixel& c, float age, float alpha)
   {
     if (next == nullptr) {
       auto xi = static_cast<int>(x);
@@ -54,7 +54,7 @@ protected:
 
 template <int W>
 class FilterRaw : public Filter<W> {
-  void plot(Canvas& canvas, double x, double y, const Pixel& c, double age, double alpha) {
+  void plot(Canvas& canvas, float x, float y, const Pixel& c, float age, float alpha) {
     this->pass(canvas, x, y, c, age, alpha);
   }
 };
@@ -64,13 +64,13 @@ class FilterAntiAlias : public Filter<W> {
 public:
   FilterAntiAlias() {}
 
-  void plot(Canvas& canvas, double x, double y, const Pixel& c, double age, double alpha) {
-    double x_i = 0;
-    double x_m = modf(x, &x_i);
-    double y_i = 0;
-    double y_m = modf(y, &y_i);
+  void plot(Canvas& canvas, float x, float y, const Pixel& c, float age, float alpha) {
+    float x_i = 0;
+    float x_m = modf(x, &x_i);
+    float y_i = 0;
+    float y_m = modf(y, &y_i);
 
-    double v = (1 - x_m) * (1 - y_m);
+    float v = (1 - x_m) * (1 - y_m);
     this->pass(canvas, x_i, y_i, c, age, alpha * v);
 
     v = x_m * (1 - y_m);
@@ -95,7 +95,7 @@ public:
     num_pixels(0)
   {}
 
-  void plot(Canvas& canvas, double x, double y, const Pixel& color, double age, double alpha) {
+  void plot(Canvas& canvas, float x, float y, const Pixel& color, float age, float alpha) {
     if (age >= 0) {
       if (num_pixels < MAX_PIXELS) {
         ttls[num_pixels++] = { static_cast<float>(x), static_cast<float>(y), static_cast<float>(lifetime - age) };
@@ -123,7 +123,7 @@ public:
     Filter<W>::decay();
   }
 
-  void trail(Canvas& canvas, TrailFn trailFn, double alpha) {
+  void trail(Canvas& canvas, TrailFn trailFn, float alpha) {
     for (int i = 0; i < num_pixels; ++i) {
       auto color = trailFn(ttls[i].x, ttls[i].y, 1 - (ttls[i].ttl / lifetime));
       this->pass(canvas, ttls[i].x, ttls[i].y, color, lifetime - ttls[i].ttl, alpha);
@@ -152,7 +152,7 @@ public:
   {
   }
 
-  void plot(Canvas& canvas, double x, double y, const Pixel& color, double age, double alpha) {
+  void plot(Canvas& canvas, float x, float y, const Pixel& color, float age, float alpha) {
     CRGB r(color.r, 0, 0);
     CRGB g(0, color.g, 0);
     CRGB b(0, 0, color.b);
@@ -170,7 +170,7 @@ public:
     orientation(orientation)
   {}
 
-  void plot(Canvas& canvas, double x, double y, const Pixel& color, double age, double alpha) {
+  void plot(Canvas& canvas, float x, float y, const Pixel& color, float age, float alpha) {
     auto v = pixel_to_vector<W>(x, y);
     auto r = vector_to_pixel<W>(orientation.orient(v));
     this->pass(canvas, r.x, r.y, color, age, alpha);
@@ -189,7 +189,7 @@ public:
     count(std::clamp(count, 1, W))
   {}
 
-  void plot(Canvas& canvas, double x, double y, const Pixel& color, double age, double alpha) {
+  void plot(Canvas& canvas, float x, float y, const Pixel& color, float age, float alpha) {
     for (int i = 0; i < W; i += W / count) {
       this->pass(canvas, wrap(x + i, W), y, color, age, alpha);
     }
