+++ util.h
@@ -3,8 +3,8 @@
 #include "3dmath.h"
 
 namespace hs {
-  double rand_dbl() {
-    return static_cast<double>(::random(0, std::numeric_limits<int32_t>::max()))
+  float rand_dbl() {
+    return static_cast<float>(::random(0, std::numeric_limits<int32_t>::max()))
       / std::numeric_limits<int32_t>::max();
   }
 
@@ -26,12 +26,12 @@ inline T wrap(T x, U m) {
   return (r >= m) ? 0 : r;
 }
 
-double shortest_distance(double a, double b, double m) {
-  double d = std::fmod(std::fmod(a - b, m) + m, m);
+float shortest_distance(float a, float b, float m) {
+  float d = std::fmod(std::fmod(a - b, m) + m, m);
   return std::min(d, m - d);
 }
 
-constexpr double fwd_distance(double a, double b, double m) {
+constexpr float fwd_distance(float a, float b, float m) {
   auto d = b - a;
   if (d < 0) {
     d += m;
